Got the bookshelf to display properly
Achieved this by adding the components MainPage, BookShelf, Book and
calling the getAll() method from BooksAPI to get a list of the books
on the backend. Main issue was using books[0], which doesn't work well
with componentDidMount becuase it resolves to undefined the first time
it renders. Modified to use map() just like in the tutorial.