SearchPage component added
So far the SearchPage can only query the backend but it doesn't find
out which books are already on the shelf or not. I want to hammer out
what to do with empty search results because it throws an error when
the search string doesn't match the given predefined strings. Also I'm
getting 429 errors in the console, not sure if I need to rerun webpack.