SearchPage functionality done
The SearchPage component now displays correct shelf that the books are
on and also allows for adding books to the shelf. Several notes: I had
to modify the changeShelf method to do a shallow copy of the book
because I intially made the error of using a reference (not good).
Also, the bookshelf now filters for shelf names that are contained in
the prop shelf name, allowing me to make a mega shelf containing books
from all shelves.