final draft
Major fix: the state for the search page wasn't working because the
sheves should be rendering from the passed state and not the state in
the component. I don't know if this is not the intention of the rubric
when it says that the book has to have the same state in the main and
search pages.

One more quirk: in order to avoid 403 error from the API, I'm forcing
the search query to not be empty string. Maybe there's a better way to
handle this.