Project Finished
Solved the empty string search input with a ternary operator, hope
it's stylistically correct. One quirk of the final product is all of
the main logic for shelf changing and searching is in App.js and is
being passed as props. All of the books and search results are also
being stored as a state in App.js and being passed as props. I used to
have a state that determined if the search result was invalid or not,
but now I just have an empty shelf message.