Ready to submit!
I found out the 409 error was coming from a Chrome extension that has
nothing to do with the app so we're good! One more bug fix: the
changeShelf function in App.js was sending books to the search page
because newBooks === [] is not a valid way to check if the array is
empty.